check_tx_input rework + fee_algo tweak and mixin refactor

+++ src/cryptonote_config.h
@@ -274,7 +274,7 @@ namespace config
    namespace tx_settings
    {
      uint_fast8_t const ARQMA_TX_CONFIRMATIONS_REQUIRED = 4; // How many blocks are needed to confirm transaction sent.
-     uint_fast8_t const ARQMA_TX_LOCK_SECONDS = 360;  // Transaction lock stated in seconds related to time-based per_output_unlock.
+     uint_fast16_t const ARQMA_TX_LOCK_SECONDS = 360;  // Transaction lock stated in seconds related to time-based per_output_unlock.
      uint_fast8_t const CURRENT_TX_VERSION = 3; // Current Transaction Version Valid on Arq-Net
      uint64_t const TRANSACTION_SIZE_LIMIT = 48 * 1024; // I did set it to 48kB for now but it need to be verified.
      uint64_t const MAX_TRANSACTIONS_IN_BLOCK = 1024; // Maximum allowed transactions in One Block
